+++ cookbooks/apple_itunes/metadata.rb
@@ -1,6 +1,6 @@
 maintainer       "Wesleyan University"
 maintainer_email "rchristensen@wesleyan.edu"
 license          "All rights reserved"
-description      "Installs iTunes 10.7"
+description      "Installs iTunes 11"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.1.0"
+version          "0.2.0"