+++ cookbooks/apple_safari/recipes/mac.rb
@@ -11,8 +11,8 @@ dmg_package "safari-6.0.2.dmg" do
   app "Safari6.0.2Mountain"
   volumes_dir "Safari"
   dmg_name "safari-6.0.2"
-  source "http://baratheon.class.wesleyan.edu/mac_os_x-10.8/apple_safari-6/safari-6.0.2.dmg"
-  checksum "51da3cd7ff88ce1c4b9840c6abcba48097ed54330c14639b0a59ba3f1006e138"
+  source "http://baratheon.class.wesleyan.edu/os_x-10.8/apple_safari/safari-6.0.2.dmg"
+  checksum "b77f25504e9a8a1231bd54e67ad00fd04a4be05681d66e571c5bf13576d5ecba"
   action :install
   type "pkg"
   package_id "com.apple.pkg.Safari6.0.2Mountain"