+++ cookbooks/microsoft_silverlight/recipes/mac.rb
@@ -10,7 +10,7 @@ dmg_package "Silverlight.dmg" do
   app "Silverlight"
   volumes_dir "Silverlight"
   dmg_name "Silverlight"
-  source "http://baratheon.class.wesleyan.edu/mac_os_x-10.8/microsoft_silverlight-5/Silverlight.dmg"
+  source "http://baratheon.class.wesleyan.edu/os_x-10.8/microsoft_silverlight/Silverlight.dmg"
   checksum "aac8a67e234e93f56751598cb67da6e58ebb9aaa93469dd038c9717f2490e674"
   action :install
   type "pkg"