+++ cookbooks/rstudio/recipes/mac.rb
@@ -6,12 +6,12 @@
 #
 # All rights reserved - Do Not Redistribute
 #
-dmg_package "RStudio-0.96.331.dmg" do
+dmg_package "RStudio-0.97.248.dmg" do
   app "RStudio"
-  volumes_dir "RStudio-0.96.331"
-  dmg_name "RStudio-0.96.331"
-  source "http://baratheon.class.wesleyan.edu/mac_os_x-10.8/rstudio-0/RStudio-0.96.331.dmg"
-  checksum "118bf97dcc4640b77add79d97df8c3c9a0eadb1bfe4d14403b8b5bda03755121"
+  volumes_dir "RStudio-0.97.248"
+  dmg_name "RStudio-0.97.248"
+  source "http://baratheon.class.wesleyan.edu/os_x-10.8/rstudio/RStudio-0.97.248.dmg"
+  checksum "4798f664ebf7d24f9c211001b0673e96d1a54e0b0f87217e9665a8427ceda71c"
   action :install
   type "app"
 end