+++ cookbooks/unarchiver/recipes/mac.rb
@@ -5,12 +5,12 @@
 # Copyright 2012, Wesleyan University
 # All rights reserved - Do Not Redistribute
 #
-dmg_package "TheUnarchiver3.4" do
+dmg_package "theunarchiver-3.6.dmg" do
   app "The Unarchiver"
-  volumes_dir "TheUnarchiver3.4"
-  dmg_name "TheUnarchiver3.4"
-  source "http://baratheon.class.wesleyan.edu/mac_os_x-10.8/unarchiver-3/TheUnarchiver3.4.dmg"
-  checksum "780e168be15d4a28061c7fa223f72849f1b42e47b873f48de9c644c16733a8d8"
+  volumes_dir "The Unarchiver"
+  dmg_name "theunarchiver-3.6"
+  source "http://baratheon.class.wesleyan.edu/os_x-10.8/unarchiver/theunarchiver-3.6.dmg"
+  checksum "65a47029ba7c37668b414faded859db691dc2969c47c07eb97a9626edaf5a3c0"
   action :install
   type "app"
   package_id "cx.c3.theunarchiver"